~ inspect for record and data
